+++ src/main/java/plus/dragons/createdragonsplus/integration/ModIntegration.java
@@ -29,7 +29,8 @@ public enum ModIntegration {
     IMMERSIVE_ENGINEERING(Constants.IMMERSIVE_ENGINEERING),
     DYE_DEPOT(Constants.DYE_DEPOT),
     DYENAMICS(Constants.DYENAMICS),
-    ARTS_AND_CRAFTS(Constants.ARTS_AND_CRAFTS);
+    ARTS_AND_CRAFTS(Constants.ARTS_AND_CRAFTS),
+    AETHER(Constants.AETHER);
 
     private final String id;
 
@@ -61,5 +62,6 @@ public enum ModIntegration {
         public static final String DYE_DEPOT = "dye_depot";
         public static final String DYENAMICS = "dyenamics";
         public static final String ARTS_AND_CRAFTS = "arts_and_crafts";
+        public static final String AETHER = "aether";
     }
 }